FIX: Only connect template_iterator_wf if it exists
